requirements.txt
Django==5.0.6
djangorestframework==3.15.1
django-filter==23.5
psycopg2-binary==2.9.9
Pillow==10.2.0 